requirements.txt
Django==6.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
django-cors-headers==4.3.1
django-filter==24.1
